refactor(common): 패키지명 변경
