+++ src/main/java/nextstep/subway/common/domain/FindFailedException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.global.domain;
+package nextstep.subway.common.domain;
 
 public class FindFailedException extends RuntimeException {
