+++ src/main/java/nextstep/subway/line/domain/SectionRemoveFailedException.java
@@ -1,6 +1,6 @@
 package nextstep.subway.line.domain;
 
-import nextstep.subway.global.domain.BusinessException;
+import nextstep.subway.common.domain.BusinessException;
 
 public class SectionRemoveFailedException extends BusinessException {
